New profile and orders (#41)

+++ palacepetz/Models/User/DtoCards.cs
@@ -7,7 +7,7 @@ namespace palacepetz.Models.Cards
 {
     public class DtoCards
     {
-        public int cd_card { get; set; }
+        public long cd_card { get; set; }
         public int id_user { get; set; }
         public int length { get; set; }
         public string flag_card { get; set; }
